shell: Add common submodule for git and make
